requirements.txt
astroid==1.5.3
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
idna==2.5
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
requests==2.18.3
six==1.10.0
twx.botapi==3.1.1
urllib3==1.22
wrapt==1.10.10
